+++ CHANGELOG.md
@@ -19,3 +19,4 @@
 - Add chips component into elements folder, first version with functions but without table updating
 - Add dialog component into elements folder to be shown when click on any of the superheroes to show his/her info
 - Add grid in the header
+- Bugfix: clear chips input when an element has been selected